GPS_Funcsの追加
