+++ raspberry_pi/lib/RoverFuncs/__init__.py
@@ -1,1 +1,2 @@
-from .core import *
+from BME_Judge import*
+from GPS_Funcs import*
