+++ raspberry_pi/main.py
@@ -102,45 +102,6 @@ goal_lat, goal_long = 40.1427210, 139.9874711 #本番用
 
 drift = -1.032555
 
-def gps_get():
-    global to_goal, rotation, pre
-    flag = 0
-    while 1:
-        now = GPS.lat_long_measurement()
-        if now[0] != None and now[1] != None:
-            to_goal[0] = GPS.convert_lat_long_to_r_theta(now[0],now[1],goal_lat,goal_long)[0]
-            print(to_goal[0])
-
-            if flag == 0 and GPS.convert_lat_long_to_r_theta(pre[0], pre[1], now[0], now[1])[0] >= forward_dis:
-                lock.acquire()
-                to_goal[1] = -math.degrees(GPS.convert_lat_long_to_r_theta(now[0],now[1],goal_lat,goal_long)[1])
-                rotation = -math.degrees(GPS.convert_lat_long_to_r_theta(pre[0], pre[1], now[0], now[1])[1])
-                lock.release()
-                print("count!!!", now)
-                pre = now
-                flag = 1
-            if to_goal[0] < cam_dis:
-                break
-
-def gyro_get():
-    global to_goal, rotation, dL, dR, m
-    pt = time.time()
-    while 1:
-        #dutyLを変える
-        gyro = mpu.get_gyro_data_lsb()[2] + drift
-        nt = time.time()
-        dt = nt - pt
-        pt = nt
-        rotation += gyro * dt
-        m = p.update_pid(to_goal[1] , rotation, dt)
-        m1 = min([max([m, -1]), 1])
-        dL, dR = 75000 + 12500 * (1 - m1), 75000 - 12500 * (1 + m1)
-        print([m, rotation, to_goal[1] - rotation])
-        time.sleep(0.01)
-
-        if to_goal[0] < cam_dis:
-            break
-
 while 1:
     pre = GPS.lat_long_measurement()
     if pre[0] != None:
